requirements.txt
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.13.1
flask==3.0.0
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
flask-sqlalchemy==3.1.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.2
pycparser==2.21
PyMySQL==1.1.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
requests-file==1.5.1
six==1.16.0
SQLAlchemy==2.0.23
tldextract==5.1.1
typing-extensions==4.9.0
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
